requirements.txt
Django
djangorestframework
drf-yasg
bs4
requests